Fixed result viewing, added APIv2 and request proxying
