+++ lib/resolvers/index.js
@@ -80,7 +80,7 @@ function main (url, opts, callback) {
     // resolverData.canonical = canonical;
     // return callback(resolverData);
 
-    console.log(resolverData);
+    // console.log(resolverData);
     return callback(canonical);
   });
 }
