+++ public/css/app.css
@@ -1,12 +1,15 @@
 h1 {font-size: 30px; font-weight: 300;}
+h2 {font-size: 18px; font-weight: 300;}
 html {position: relative; min-height: 100%;}
 body {margin-bottom: 40px;}
+h1 a, h1 a:hover, h1 a:focus {color: black; text-decoration: none}
 
 .footer {
   position: absolute; bottom: 0; width: 100%; height: 40px; background-color: #f5f5f5;
 }
 
-.footer .container {line-height: 40px;}
+.footer .container {line-height: 40px}
+.footer, .footer span {color: #AAA}
 .footer a {color: #888; margin: 0 5px;}
 .footer a:hover, .footer a:focus, .footer a:active {color: #333;}
