+++ views/layout.html
@@ -4,7 +4,8 @@
 <body>
 
   <div class="text-center">
-    <h1><strong>URL</strong>Gent</h1>
+    <h1><a href="/"><strong>URL</strong>Gent</a></h1>
+    <h2>Because privacy matters</h2>
   </div>
 
   <br /><br />
