+++ views/pages/api.html
@@ -1,13 +1,15 @@
-<h2>API</h2>
+<p><i>Throttling may apply.</i></p>
 
-<p>This API provides just one endpoint at the moment as that's all I need for now.</p>
-<p>You can connect to it using regular HTTP GETs</p>
-<p>Example: <code>curl http://urlgent.com/api/v1/resolve?url=http://example.com&type=html</code>.</p>
-<p>Throttling may apply.</p>
+<h3>API v1</h3>
+
+<p>In API v1 the whole lifecycle is packed into a single request.
+  This may be preferred in some circumstances.</p>
+
+<br />
 
 <table class="table" width="100%">
   <tr>
-    <td>
+    <td width="380">
       <span class="label label-primary">GET</span>
       <code>/api/v1/resolve</code>
       <br /><br />
@@ -21,7 +23,82 @@
       </ul>
     </td>
     <td>
-      Fetches information about given URL
+      Fetches information about given URL.
+      <br /><br />
+      <strong>Example</strong><br />
+      <code>curl http://urlgent.com/api/v1/resolve?url=https:\/\/www.youtube.com\/watch\?v=C0DPdy98e4c&type=media</code>
+    </td>
+  </tr>
+</table>
+
+<br />
+
+<h3>API v2</h3>
+
+<p>In API v2 uses a job concept and breaks the process into three distinct parts.</p>
+<ol>
+  <li>Pass <code>url</code> and <code>type</code> to <code>create</code> endpoint.</li>
+  <li>Look it up the results through <code>get</code> endpoint.</li>
+  <li>Use the provided URL directly or through <code>stream</code> endpoint to prevent tracking when rendering embedded video.</li>
+</ol>
+
+<br />
+
+<table class="table" width="100%">
+  <tr>
+    <td width="380">
+      <span class="label label-primary">GET</span>
+      <code>/api/v2/create</code>
+      <br /><br />
+
+      <strong>Params:</strong>
+      <br />
+
+      <ul>
+        <li>url</li>
+        <li>type ('html' or 'media')</li>
+      </ul>
+    </td>
+    <td>
+      Creates a resolution request. Returns request ID.
+    </td>
+  </tr>
+
+  <tr>
+    <td>
+      <span class="label label-primary">GET</span>
+      <code>/api/v2/get/:id</code>
+      <br /><br />
+
+      <strong>Params:</strong>
+      <br />
+
+      <ul>
+        <li>id - request ID from <code>/create</code> endpoint</li>
+      </ul>
+    </td>
+    <td>
+      Retrieves the information about URL provided when creating request.
+    </td>
+  </tr>
+
+  <tr>
+    <td>
+      <span class="label label-primary">GET</span>
+      <code>/api/v2/stream/:id/:type/:nicename</code>
+      <br /><br />
+
+      <strong>Params:</strong>
+      <br />
+
+      <ul>
+        <li>id</li>
+        <li>type ('audio' or 'video')</li>
+        <li>nicename (optional)</li>
+      </ul>
+    </td>
+    <td>
+      Pipes the audio or video file to users browser to maintain privacy
     </td>
   </tr>
 </table>
