+++ views/pages/home.html
@@ -14,14 +14,5 @@
       </div>
     </form>
 
-    <div class="text-center">
-      <br /><br />
-      {{!
-      <a href="https://chrome.google.com/webstore/detail/negdabnedmjkiognhhikoggfbbmpjjmj" target="_blank">
-        Get URLGent Chrome Extension
-      </a>
-      }}
-    </div>
-
   </div>
 </div>
