+++ views/partials/footer.html
@@ -1,5 +1,7 @@
 <footer class="footer">
   <div class="container text-center">
+    <span>You are not being tracked.</span>
+    &nbsp; | &nbsp;
     <a href="/">Home</a>
     <a href="/api">API</a>
     <a href="/stats">Stats</a>
